# Eliminate (almost) all invalidations by disabling world splitting