Let's see if we can keep runtime performance while getting rid of a major cause of compile time performance issues.

- **Set max methods to 1 (mostly disable world splitting)**
- **Also disable word splitting when the matches are not fully covered (regardless of number of methods) [Using Claude]**