@topolarity 2025-07-25
Oof, 17 failures for `--trim` hello, world:
```julia
Trim verify finished with 17 errors, 0 warnings.
```

A lot of that would be improved dramatically by making sure we world-split for any anonymous function types (e.g. closures + inner functions for args expansion). Anonymity means that these are pretty much never extended so they should be treated as `final`