@LilithHafner 2025-07-25
We need a good heuristic for 

> Will this call see no more new methods?

---

Candidates:

1) Less than 4 matching methods (release, master)
2) Exactly one matching method whose type signature fully covers the possible argument types (pr)
3) Only if the argument types are concrete (even more conservative than pr, eliminates pretty much all non-piracy invalidations)
4) If the function is anonymous
5) If the function has been marked `final`
6) If the callsite has been marked `@worldsplit` (or maybe we want to call it `@static_dispatch`)
7) `2 | 4 | 5 | 6`
8) `3 | 4 | 5 | 6`