@nsajko 2025-07-25
Furthermore, as I discuss there, setting `max_methods` to one is not appropriate for a wide class of functions, those which take types as arguments and need to have a separate method for the bottom type, because the bottom type subtypes each type so the method will appear in method queries even when it's not strictly relevant.