@LilithHafner 2025-07-25
@nsajko, is `f`/`g` the "wide class of functions" you are referring to? Apparently there is a union splitting optimization for types that this PR fails to disable* which should make that a non-issue.

```
x@fedora:~/.julia/dev/julia$ ./julia --startup=no
               _
   _       _ _(_)_     |  Documentation: https://docs.julialang.org
  (_)     | (_) (_)    |
   _ _   _| |_  __ _   |  Type "?" for help, "]?" for Pkg help.
  | | | | | | |/ _' |  |
  | | |_| | | | (_| |  |  Version 1.13.0-DEV.895 (2025-07-23)
 _/ |\__'_|_|_|\__'_|  |  lh/max-methods-1/1450d944c3* (fork: 1 commits, 1 day)
|__/                   |

julia> f(::Type{Int}) = 1
f (generic function with 1 method)

julia> f(::Type{Union{}}) = 0.0
f (generic function with 2 methods)

julia> g(x) = f(x[1])
g (generic function with 1 method)

julia> Base.infer_return_type(g, Tuple{Vector{Union{Type{Union{}}, Type{Int}}}})
Union{Float64, Int64}

julia> f2(x::Int) = 1
f2 (generic function with 1 method)

julia> f2(x::Float64) = :a
f2 (generic function with 2 methods)

julia> g2(x) = f2(x[1])
g2 (generic function with 1 method)

julia> Base.infer_return_type(g, Tuple{Vector{Union{Int, Float64}}})
Any
```

*It'd be reassuring if someone could point out where that optimization takes place.