@nanosoldier 2025-07-26
The package evaluation job [you requested](https://github.com/JuliaLang/julia/pull/59091#issuecomment-3115440942) has completed - possible new issues were detected.
The [**full report**](https://s3.amazonaws.com/julialang-reports/nanosoldier/pkgeval/by_hash/efa4faf_vs_96b5b2c/report.html) is available.

<details><summary>Report summary</summary>

## ❗ Packages that crashed

**6 packages crashed only on the current version.**

 - An internal error was encountered: 5 packages
 - GC corruption was detected: 1 packages

6 packages crashed on the previous version too.

## ✖ Packages that failed

**159 packages failed only on the current version.**

 - Package fails to precompile: 28 packages
 - Illegal method overwrites during precompilation: 1 packages
 - Package has test failures: 23 packages
 - Package tests unexpectedly errored: 98 packages
 - Package is using an unknown package: 1 packages
 - Networking-related issues were detected: 1 packages
 - Tests became inactive: 2 packages
 - Test duration exceeded the time limit: 5 packages

1195 packages failed on the previous version too.

## ✔ Packages that passed tests

**31 packages passed tests only on the current version.**

 - Other: 31 packages

5274 packages passed tests on the previous version too.

## ~ Packages that at least loaded

**26 packages successfully loaded only on the current version.**

 - Other: 26 packages

3176 packages successfully loaded on the previous version too.

## ➖ Packages that were skipped altogether

907 packages were skipped on the previous version too.



</details>