@LilithHafner 2025-07-26
Just looked through a nonrandom sample of 7 of the new failures and 6 of them were literally testing inference results or performance or display based derivatives of inference results and [one](https://s3.amazonaws.com/julialang-reports/nanosoldier/pkgeval/by_hash/efa4faf_vs_96b5b2c/RegistryTools.primary.log) had a MethodError from a `Vector{Any}` (also dependent on inference results, but at least a semantic failure). This final category is the highest priority to fix prior to merging.