@LilithHafner 2025-07-26
Yes, obviously. This PR doesn't hurt runtime of concretely typed code, though.