@LilithHafner 2025-07-29
Here's the JuliaInterpreter.jl benchmarks
```
julia> judge(pr, nightly).benchmarkgroup
7-element BenchmarkTools.BenchmarkGroup:
  tags: []
  "recursive other 1_000" => TrialJudgement(+32.02% => regression)
  "recursive self 1_000" => TrialJudgement(+33.79% => regression)
  "throw long 1_000" => TrialJudgement(-34.07% => improvement)
  "long function 5_000" => TrialJudgement(+38.60% => regression)
  "tight loop 10_000" => TrialJudgement(+23.62% => regression)
  "ccall library" => TrialJudgement(+4.85% => invariant)
  "ccall ptr" => TrialJudgement(+21.08% => regression)
```