@timholy 2025-07-29
It's definitely a hit, but not as bad as I feared.