@oscardssmith 2025-11-18
For some TTFX benchmarks vs master: TLDR: 2.3x faster `using OrdinaryDiffEq`, ~30% faster precompilation of OrdinaryDiffEq and all deps, ~30% faster building of julia.

master:
```
Compiling the compiler. This may take several minutes ...
Base.Compiler ──── 551.44097995758057 seconds

JuliaSyntax/src/JuliaSyntax.jl
Base  ────────── 56.298068 seconds
FileWatching  ──  8.012995 seconds
Libdl  ─────────  0.003554 seconds
Artifacts  ─────  0.413654 seconds
SHA  ───────────  0.276992 seconds
Sockets  ───────  0.366529 seconds
LinearAlgebra  ─  9.732326 seconds
Random  ────────  0.962642 seconds
Stdlibs total  ─ 19.776286 seconds
Sysimage built. Summary:
Base ────────  56.298068 seconds 74.004%
Stdlibs ─────  19.776286 seconds 25.996%
Total ───────  76.074388 seconds

Outputting sysimage file...
Output ────── 404.520990 seconds

Precompiling packages finished. #stdlib
  106 dependency configurations successfully precompiled in 515 seconds
  
Precompiling OrdinaryDiffEq finished.
  207 dependencies successfully precompiled in 500 seconds. 35 already precompiled.

julia> @time using OrdinaryDiffEq
  8.371604 seconds (4.56 M allocations: 288.796 MiB, 1.35% gc time, 1.31% compilation time)

julia> f(u,p,t) = u;
julia> @time prob = ODEProblem(f, rand(2), (0.0, 1.0));
  0.092285 seconds (132.07 k allocations: 7.774 MiB, 95.87% compilation time: 87% of which was recompilation)

julia> @time solve(prob);
  7.677844 seconds (12.05 M allocations: 634.319 MiB, 1.13% gc time, 100.00% compilation time)
```
PR:
```
Compiling the compiler. This may take several minutes ...
Base.Compiler ──── 390.62850403785706 seconds

JuliaSyntax/src/JuliaSyntax.jl
Base  ────────── 45.306912 seconds
FileWatching  ──  6.154402 seconds
Libdl  ─────────  0.001715 seconds
Artifacts  ─────  0.483174 seconds
SHA  ───────────  0.220964 seconds
Sockets  ───────  0.278768 seconds
LinearAlgebra  ─  7.737392 seconds
Random  ────────  0.770357 seconds
Stdlibs total  ─ 15.653594 seconds
Sysimage built. Summary:
Base ────────  45.306912 seconds 74.3217%
Stdlibs ─────  15.653594 seconds 25.6782%
Total ───────  60.960523 seconds

Outputting sysimage file...
Output ────── 361.450666 seconds

Precompiling packages finished. #stdlib
  106 dependency configurations successfully precompiled in 415 seconds

Precompiling OrdinaryDiffEq finished.
  207 dependencies successfully precompiled in 377 seconds. 35 already precompiled.

julia> @time using OrdinaryDiffEq
  3.632475 seconds (3.12 M allocations: 178.062 MiB, 1.47% gc time, 2.36% compilation time)

julia> f(u,p,t) = u;
julia> @time prob = ODEProblem(f, rand(2), (0.0, 1.0));
  0.019137 seconds (50.34 k allocations: 3.119 MiB, 79.69% compilation time)

julia> @time solve(prob);
  7.684229 seconds (11.42 M allocations: 616.780 MiB, 2.89% gc time, 100.00% compilation time)
```