@jakobnissen 2025-11-23
A few more observations, when using Turing.jl, another dependency-heavy package, and comparing this PR versus 1.12.1:

* The number of unique invalidations when loading Turing falls by 83% from 2496 to 423. This implies there are still a bunch of invalidations remaining even on this PR. A brief scan over some of the invalidations tell me that these stem from world-splitting callsites that previously had only a single matching method. That suggests there is a substantial difference between setting MAX_METHODS to 1 (what this PR currently does) and disabling world-splitting for generic, extensible code altogether, i.e. option 8 in [this comment](https://github.com/JuliaLang/julia/pull/59091#issuecomment-3117438105).
* This PR drops `using` time from 1.7s to 1.4s. It's not clear to me whether this drop in load time, from fewer invalidations in Pkg code is generalizable to fewer invalidations in non-Pkg code. On one hand, Pkg is highly type unstable and therefore more prone to invalidations. On the other hand, package loading has received a LOT of developer attention to reduce invalidations.
* Loading + TTFX for the Turing README example drops from 15.6s to 13.3s.
* This PR increases the amount of dynamic dispatched when loading the package (measured using `--trace-dispatch`  from 330 to 821. On one hand this shows that world splitting _works_, by statically resolving ~500 call sites. On the other, it shows that the observed package loading time drop from 1.7 to 1.4 is an underestimation, since quite a lot of these 500 now-unresolved call sites could presumably be written to be type stable and therefore be faster.

Overall, my conclusion is:
* The immediate benefits of this PR are minor
* This PR elimintates the majority of world splitting, but nowhere near "almost all" as the title states.
* Merging this PR has some short term pain by making significantly more call sites dynamic.

I'm in favor, but I advocate for _actually_ eliminate almost all world splitting by going for [option 8](https://github.com/JuliaLang/julia/pull/59091#issuecomment-3117438105) or similar.